refactor: decompose skill_creator + rules_creator god-modules
skill_creator.py: 1190 → 824 LOC (−31%)
- Extract SkillDocLoader → generator/skill_doc_loader.py
- Extract SkillMetadataBuilder → generator/skill_metadata_builder.py

rules_creator.py: 864 → 622 LOC (−28%)
- Extract RulesGitMiner → generator/rules_git_miner.py
- Extract RulesContentRenderer → generator/rules_renderer.py

Consolidate quality validators:
- SkillQualityValidator + RulesQualityValidator → generator/quality_validators.py
- Removes skill_quality_validator.py and rules_quality_validator.py

Docs:
- docs/KNOWN-ISSUES.md: 4 bugs found in prg design/plan pipeline
- docs/DESIGN-skill-creator-refactor.md: decomposition design record
- CHANGELOG.md, docs/architecture.md: updated

All public APIs unchanged. 67 targeted tests pass.
