+++ generator/rules_renderer.py
@@ -1,0 +1,111 @@
+"""
+Rules Content Renderer
+======================
+
+Renders the final rules.md content from metadata and rule collections.
+Extracted from CoworkRulesCreator to keep each module focused.
+
+Also owns append_mandatory_anti_patterns — kept here for cohesion since
+it is purely a rendering concern (appending to generated text).
+"""
+
+from pathlib import Path
+from typing import Dict, List, Set, TYPE_CHECKING
+
+import yaml
+
+if TYPE_CHECKING:
+    from generator.rules_creator import Rule, RulesMetadata
+
+
+def append_mandatory_anti_patterns(generated_rules: str, framework: str = "") -> str:
+    """Forces the LLM output to include explicit negative constraints."""
+    section = "\n## \U0001f6ab Critical Anti-Patterns (NEVER DO THIS)\n\n"
+    section += "- NEVER run destructive commands (rm -rf, drop table, etc.) without explicit user confirmation.\n"
+    section += "- NEVER generate placeholder comments like 'Implement logic here' - write the full code.\n"
+
+    if framework and "react" in framework.lower():
+        section += "- NEVER mutate React state directly; always use the setter functions.\n"
+
+    return generated_rules + section + "\n"
+
+
+class RulesContentRenderer:
+    """Renders rules.md content from metadata and rule collections."""
+
+    def render(
+        self,
+        metadata: "RulesMetadata",
+        rules_by_category: Dict[str, List["Rule"]],
+        readme_content: str = "",
+    ) -> str:
+        """Generate complete rules.md content."""
+        frontmatter = {
+            "project": metadata.project_name,
+            "tech_stack": metadata.tech_stack,
+            "priority_rules": metadata.priority_areas,
+            "project_type": metadata.project_type,
+            "version": "2.0",
+            "generated": "cowork-powered",
+        }
+        yaml_str = yaml.dump(frontmatter, sort_keys=False, allow_unicode=True)
+
+        content = (
+            f"---\n{yaml_str}---\n\n"
+            f"# {metadata.project_name} - Coding Rules\n\n"
+            f"**Generated by Cowork-Powered Rules Creator** \U0001f916\n\n"
+            f"## \U0001f4cb Priority Areas\n\n"
+            f"{self._format_priority_areas(metadata.priority_areas)}\n\n"
+            f"## \U0001f4bb Coding Standards\n\n"
+        )
+
+        all_rules_flat = [rule for rules in rules_by_category.values() for rule in rules]
+        shown_in_priority: Set[str] = set()
+
+        for priority in ["High", "Medium", "Low"]:
+            priority_rules = [r for r in all_rules_flat if r.priority == priority]
+            if priority_rules:
+                content += f"\n### {priority} Priority\n\n"
+                for rule in priority_rules:
+                    content += f"- \u2705 **{rule.content}**\n"
+                    shown_in_priority.add(rule.content)
+
+        content += "\n## \U0001f4c2 Rules by Category\n\n"
+        priority_emoji = {"High": "\U0001f534", "Medium": "\U0001f7e1", "Low": "\U0001f7e2"}
+        for category, rules in sorted(rules_by_category.items()):
+            unique_rules = [r for r in rules if r.content not in shown_in_priority]
+            if not unique_rules:
+                continue
+            content += f"\n### {category}\n\n"
+            for rule in unique_rules:
+                emoji = priority_emoji.get(rule.priority, "\U0001f7e2")
+                content += f"- {emoji} {rule.content}\n"
+
+        content += "\n## \U0001f6e0\ufe0f Tech Stack\n\n"
+        for tech in metadata.tech_stack:
+            content += f"- {tech}\n"
+
+        if metadata.detected_signals:
+            content += "\n## \U0001f3d7\ufe0f Project Structure\n\n"
+            for signal in metadata.detected_signals:
+                content += f"- `{signal}`\n"
+
+        content += "\n---\n"
+        content += "*Generated by Cowork-Powered PRG Rules Creator*\n"
+        total_rules = sum(len(rules) for rules in rules_by_category.values())
+        content += (
+            f"*Tech Stack: {len(metadata.tech_stack)} | "
+            f"Rules: {total_rules} | "
+            f"Quality: Cowork-level*\n"
+        )
+
+        framework_hint = "react" if "react" in metadata.tech_stack else ""
+        content = append_mandatory_anti_patterns(content, framework_hint)
+
+        return content
+
+    def _format_priority_areas(self, areas: List[str]) -> str:
+        """Format priority areas as markdown list."""
+        if not areas:
+            return "- No specific priority areas detected"
+        return "\n".join(f"- **{area.replace('_', ' ').title()}**" for area in areas)
